feat: delete node-lifecycle.js + node-orchestration.js, clean up ActionContext
Migrate node-lifecycle ownership side-effects (cancel trace on security-monitor owned, disable ICE on resident owned) to game-ctx.js as module-level listener. Delete node-lifecycle.js and node-orchestration.js. Strip stale stubs from ActionContext typedef and action-context.js. Remove initNodeLifecycle calls from all entry points.
