+++ js/core/types.js
@@ -192,24 +192,17 @@
  * Dependency-injection context passed to action execute() functions.
  * main.js constructs one instance at init, wiring each field to the
  * corresponding state mutator. Tests can pass mock contexts.
+ * Action context for the dispatcher — only contains actions not handled by the
+ * graph's trait-based action system. Most game actions (probe, exploit, read,
+ * loot, reboot, reconfigure) are now graph-native via timed-action operators.
  * @typedef {{
- *   getState:      () => GameState,
- *   selectNode:    (nodeId: string) => void,
- *   deselectNode:  () => void,
- *   startProbe:    (nodeId: string) => void,
- *   cancelProbe:   () => void,
- *   startExploit:  (nodeId: string, exploitId: string) => void,
- *   cancelExploit: () => void,
- *   startRead:     (nodeId: string) => void,
- *   cancelRead:    () => void,
- *   startLoot:     (nodeId: string) => void,
- *   cancelLoot:    () => void,
- *   ejectIce:         () => void,
- *   rebootNode:       (nodeId: string) => void,
- *   jackOut:          () => void,
- *   reconfigureNode:     (nodeId: string) => void,
- *   cancelTrace:         () => void,
- *   openDarknetsStore:   () => void,
+ *   getState:          () => GameState,
+ *   selectNode:        (nodeId: string) => void,
+ *   deselectNode:      () => void,
+ *   ejectIce:          () => void,
+ *   jackOut:            () => void,
+ *   cancelTrace:       () => void,
+ *   openDarknetsStore: () => void,
  * }} ActionContext
  */
